requirements.txt
feedparser
google-generativeai
python-dotenv
pyyaml
requests
pydantic
pytest
notion-client
